+++ cas-client-core/pom.xml
@@ -35,12 +35,6 @@
         </dependency>
 
         <dependency>
-            <groupId>joda-time</groupId>
-            <artifactId>joda-time</artifactId>
-            <version>2.7</version>
-        </dependency>
-
-        <dependency>
             <groupId>xml-security</groupId>
             <artifactId>xmlsec</artifactId>
             <version>1.3.0</version>
