+++ cas-client-core/src/test/java/org/jasig/cas/client/util/CommonUtilsTests.java
@@ -21,7 +21,6 @@ package org.jasig.cas.client.util;
 import java.net.URL;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Date;
 
 import junit.framework.TestCase;
 import org.jasig.cas.client.PublicTestHttpServer;
@@ -193,10 +192,5 @@ public final class CommonUtilsTests extends TestCase {
     
     public void testUrlEncode() {
         assertEquals("this+is+a+very+special+parameter+with+%3D%25%2F", CommonUtils.urlEncode("this is a very special parameter with =%/"));
-    }
-
-    public void testParseUtcDate() {
-        final Date expected = new Date(1424437961025L);
-        assertEquals(expected, CommonUtils.parseUtcDate("2015-02-20T08:12:41.025-0500"));
     }
 }
