+++ cas-client-support-saml/pom.xml
@@ -16,7 +16,13 @@
             <artifactId>cas-client-core</artifactId>
             <version>${project.version}</version>
         </dependency>
+        <dependency>
+            <groupId>joda-time</groupId>
+            <artifactId>joda-time</artifactId>
+            <version>2.7</version>
+        </dependency>
 
+        <!-- Test dependencies -->
         <dependency>
             <groupId>org.jasig.cas.client</groupId>
             <artifactId>cas-client-core</artifactId>
